Improving testing cases

+++ backend/src/controllers/user.controller.ts
@@ -1,6 +1,8 @@
 import { FastifyRequest, FastifyReply, FastifyTypeProvider } from 'fastify'
 import { User } from '../models/user.model'
 import jwt from 'jsonwebtoken'
+import mongoose from 'mongoose'
+
 
 // ex
 
@@ -14,16 +16,12 @@ export async function getCurrent(req: FastifyRequest, reply: FastifyReply) {
 // USER ROUTES 
 
 export async function getAll(req: FastifyRequest, reply: FastifyReply) {
-    console.log('Register function called');
-    console.log('Request body:', req.body);
-    // const { id } = req.params as { id: string }
     try {
         const users = await User.find({})
         reply.code(200).send(users)
     } catch (err) {
         reply.code(500).send({ error: 'Getting all Users failed', details: err })
     }
-    // just admin
 }
 
 
@@ -31,6 +29,12 @@ export async function getAll(req: FastifyRequest, reply: FastifyReply) {
 export async function getById(req: FastifyRequest, reply: FastifyReply) {
 
     const { id } = req.params as { id: string }
+    if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+        reply.code(400).send({ error: 'ID is required' })
+        return
+    }
+
+
 
     try {
         const user = await User.findById(id)
@@ -46,9 +50,18 @@ export async function getById(req: FastifyRequest, reply: FastifyReply) {
 
 export async function updateById(req: FastifyRequest, reply: FastifyReply) {
     const { id } = req.params as { id: string }
+    if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+        reply.code(400).send({ error: 'ID is required' })
+        return
+    }
     const { name, email, password } = req.body as { name: string, email: string, password: string }
+
+    if (!name && !email && !password) {
+        reply.code(400).send({ error: 'Name, email or password are required' })
+        return
+    }
     try {
-        const user = await User.findOneAndUpdate({ '_id': id }, { 'name': name, 'email': email, 'passsword': password }, { new: true })
+        const user = await User.findOneAndUpdate({ '_id': id }, { 'name': name, 'email': email, 'password': password }, { new: true })
         if (!user) {
             reply.code(404).send({ error: 'User not found' })
             return
@@ -68,7 +81,11 @@ export async function deleteById(req: FastifyRequest, reply: FastifyReply) {
 
     const { id } = req.params as { id: string }
 
-    console.log('Request ID:', id)
+    if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+        reply.code(400).send({ error: 'ID is required' })
+        return
+    }
+
     try {
         const user = await User.findByIdAndDelete(id);
         if (!user) {
